engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-chess>=1.11.0
httpx>=0.28.0
